+++ tosho_amap/src/constants.rs
@@ -160,6 +160,9 @@ pub(crate) static MASKED_LOGIN: LazyLock<String> = LazyLock::new(|| {
 /// # Arguments
 /// * `device_type` - The device type to get the constants for.
 ///
+/// # Available device types
+/// * `1` - Android
+///
 /// # Panics
 /// Panics if the device type is invalid.
 ///