+++ tosho_rbean/src/constants.rs
@@ -111,6 +111,9 @@ pub(crate) static X_DRM_HEADER: LazyLock<String> = LazyLock::new(|| {
 /// # Arguments
 /// * `device_type` - The device type to get the constants for.
 ///
+/// # Available device types
+/// * `1` - Android
+///
 /// # Panics
 /// Panics if the device type is invalid.
 ///