+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 Starting from Rust port of the project, all changes will be put into this file.
 
-## Unreleased (git master)
+## [0.6.1] (2024-08-13)
 ### Changes
 - Fix issue with manual merging `Ctrl+C` handling not working properly